refactor: api 응용 함수들을 services 폴더로 이동
