+++ src/components/manager/user/UserBoard.tsx
@@ -5,7 +5,6 @@ import { useOverlay } from "@toss/use-overlay";
 import JSZip from "jszip";
 import { useEffect, useState } from "react";
 
-import { permanentDeleteOrdererList, reassignFolder } from "@/api";
 import { getProductList } from "@/api/products";
 import { getUserList } from "@/api/users";
 import CounselingIntakeForm from "@/components/common/CounselingIntakeForm";
@@ -24,6 +23,8 @@ import {
 import { Folder, Product, User } from "@/const";
 import { imageUrlList } from "@/recoil/user/atoms/imageUrlListState";
 import { SelectedInfoList } from "@/recoil/user/atoms/selectedInfoListState";
+import { reassignFolder } from "@/services/folders";
+import { permanentDeleteOrdererList } from "@/services/users";
 
 const StyledUserBoard = styled.div`
   display: flex;
